pyproject.toml
[tool.poetry]
name = "pdf2audio"
version = "0.1.0"
description = "Convert PDF book to audiobook using NeMo and FastPitch"
authors = ["mrmanna"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
nemo-toolkit = "^1.23.0"
torch = "^2.4.0"
pdfplumber = "^0.11.4"
librosa = "^0.10.2.post1"
matplotlib = "^3.9.2"
einops = "^0.8.0"
huggingface-hub = "^0.23.2"
transformers = "^4.44.2"
pydub = "^0.25.1"

[tool.poetry.group.dev.dependencies]
hydra-core = "^1.3.2"
pytorch-lightning = "^2.4.0"
sentencepiece = "^0.2.0"
pandas = "^2.2.2"
editdistance = "^0.8.1"
lhotse = "^1.27.0"
pyannote-audio = "^3.3.1"
webdataset = "^0.2.100"
datasets = "^2.21.0"
jiwer = "^3.0.4"
ipython = "^8.27.0"
wandb = "^0.17.8"
nemo-text-processing = "^1.1.0"
nltk = "^3.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
[tool.poetry.scripts]
start = "pdf2audio.pdf2audio:main"